// frontend/src/lib/enums.ts
export enum SetOperation {
  Union = "Union",
  Intersection = "Intersection",
  Difference = "Difference",
  SymmetricDifference = "Symmetric Difference",
  CartesianProduct = "Cartesian Product",
  PowerSet = "Power Set",
  Complement = "Complement",
}

export enum SetProperty {
  IsSubset = "isSubset",
  Cardinality = "cardinality",
  IsDisjoint = "isDisjoint",
}